requirements.txt
click==6.6
Flask==0.11.1
gevent==1.1.2
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pbr==1.10.0
psutil==4.4.2
psycogreen==1.0
psycopg2==2.6.2
py==1.4.31
Pygments==2.1.3
python-magic==0.4.12
six==1.10.0
uptime==3.0.1
urwid==1.3.1
Werkzeug==0.11.11
